cluster {runibic}	R Documentation

Calculate biclusters from sorted list of LCS scores and row indices

Description

This function search for biclusters in the input matrix. The calculations are based on the integer matrix with indexes indicating positions of j-th smallest element in each row and the results from calculations of Longest Common Subsequence between all rows in the input matrix. The paramteres of this function can be obtained from other functions provided by this package.

Usage

cluster(discreteInput, discreteInputValues, scores, geneOne, geneTwo, rowNumber,
  colNumber)

Arguments

discreteInput
an integer matrix with indices of sorted columns

discreteInputValues
an integer matrix with discrete values

scores
a numeric vector with LCS length

geneOne
a numeric vector with first row indexes from pairwise LCS calculation

geneTwo
a numeric vector with second row indexes from pairwise LCS calculation

rowNumber
a int with number of rows in the input matrix

colNumber
a int with number of columns in the input matrix

Value

a list with information of found biclusters

See Also

runibic calculateLCS unisort

Examples

A <- matrix( c(4,3,1,2,5,8,6,7,9,10,11,12),nrow=4,byrow=TRUE)
iA <- unisort(A)
lcsResults <- calculateLCS(A)
cluster(iA, A, lcsResults$lcslen, lcsResults$a, lcsResults$b, nrow(A), ncol(A))

[Package runibic version 1.14.0 Index]