runiDiscretize {runibic}	R Documentation

Discretize an input matrix

Description

This function discretizes the input matrix. runiDiscretize uses paramaters: 'div' and 'q', which are set by set_runibic_params function. The funtion returns a discrete matrix with given number of ranks based on the parameter div. In contrast to biclust::discretize the function takes into consideration the quantile parameter 'q'. When 'q' parameter is higher or equal 0.5 a simple discretization is used with equal sizes of the levels using the quantiles. If 'q' parameter is lower than 0.5 we use up(down)-regulated discretization divided into three parts.

Usage

runiDiscretize(x)

Arguments

x
a numeric matrix

Value

a discretized matrix containing integers only

See Also

set_runibic_params calculateLCS discretize

Examples

A <- replicate(10, rnorm(20))
runiDiscretize(A)

[Package runibic version 1.14.0 Index]