unisort {runibic}	R Documentation

Computing the indexes of j-th smallest values of each row

Description

This function sorts separately each row of a integer matrix and returns a matrix in which the value in i-th row and j-th column represent the index of the j-th smallest value of the i-th row.

Usage

unisort(x)

Arguments

x
a integer matrix

Value

a integer matrix with indexes indicating positions of j-th smallest element in each row

See Also

runibic calculateLCS runiDiscretize

Examples

A <- matrix(c(4, 3, 1, 2, 5, 8, 6, 7), nrow=2, byrow=TRUE)
unisort(A)

[Package runibic version 1.14.0 Index]